Security improvement for item control

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -84,11 +85,15 @@ def detail(request, product_id):
 
     return render(request, 'products/detail.html', context)
 
-    
+@login_required    
 def add_item(request):
     """
     Add items (admin)
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'limited to store owners')
+        return redirect(reverse('home'))
+
     if request.method =='POST':
         form = ItemForm(request.POST,request.FILES)
         if form.is_valid():
@@ -106,11 +111,15 @@ def add_item(request):
 
     return render(request,template,context)
 
-
+@login_required 
 def edit_item(request, product_id):
     """
     Edit item in stock
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'limited to store owners')
+        return redirect(reverse('home'))
+
     item = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form= ItemForm(request.POST, request.FILES, instance=item)
@@ -132,10 +141,15 @@ def edit_item(request, product_id):
 
     return render(request,template,context)
 
+@login_required 
 def delete_item(request, product_id):
     """
     Delete item from stock
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'limited to store owners')
+        return redirect(reverse('home'))
+
     item = get_object_or_404(Product, pk=product_id)
     item.delete()
     messages.success(request,'Item has been deleted!')
